tweak share buttons for sunny
